+++ pyccel/ast/bitwise_operators.py
@@ -37,7 +37,7 @@ class PyccelInvert(PyccelUnaryOperator):
 
     Parameters
     ----------
-    arg: PyccelAstNode
+    arg: TypedAstNode
         The argument passed to the operator
     """
     __slots__ = ()
@@ -64,9 +64,9 @@ class PyccelBitOperator(PyccelOperator):
 
     Parameters
     ----------
-    arg1: PyccelAstNode
+    arg1: TypedAstNode
         The first argument passed to the operator
-    arg2: PyccelAstNode
+    arg2: TypedAstNode
         The second argument passed to the operator
     """
     _shape = None
@@ -126,9 +126,9 @@ class PyccelRShift(PyccelBitOperator):
 
     Parameters
     ----------
-    arg1: PyccelAstNode
+    arg1: TypedAstNode
         The first argument passed to the operator
-    arg2: PyccelAstNode
+    arg2: TypedAstNode
         The second argument passed to the operator
     """
     __slots__ = ()
@@ -149,9 +149,9 @@ class PyccelLShift(PyccelBitOperator):
 
     Parameters
     ----------
-    arg1: PyccelAstNode
+    arg1: TypedAstNode
         The first argument passed to the operator
-    arg2: PyccelAstNode
+    arg2: TypedAstNode
         The second argument passed to the operator
     """
     __slots__ = ()
@@ -168,9 +168,9 @@ class PyccelBitComparisonOperator(PyccelBitOperator):
 
     Parameters
     ----------
-    arg1: PyccelAstNode
+    arg1: TypedAstNode
         The first argument passed to the operator
-    arg2: PyccelAstNode
+    arg2: TypedAstNode
         The second argument passed to the operator
     """
     __slots__ = ()
@@ -197,9 +197,9 @@ class PyccelBitXor(PyccelBitComparisonOperator):
 
     Parameters
     ----------
-    arg1: PyccelAstNode
+    arg1: TypedAstNode
         The first argument passed to the operator
-    arg2: PyccelAstNode
+    arg2: TypedAstNode
         The second argument passed to the operator
     """
     __slots__ = ()
@@ -220,9 +220,9 @@ class PyccelBitOr(PyccelBitComparisonOperator):
 
     Parameters
     ----------
-    arg1: PyccelAstNode
+    arg1: TypedAstNode
         The first argument passed to the operator
-    arg2: PyccelAstNode
+    arg2: TypedAstNode
         The second argument passed to the operator
     """
     __slots__ = ()
@@ -243,9 +243,9 @@ class PyccelBitAnd(PyccelBitComparisonOperator):
 
     Parameters
     ----------
-    arg1: PyccelAstNode
+    arg1: TypedAstNode
         The first argument passed to the operator
-    arg2: PyccelAstNode
+    arg2: TypedAstNode
         The second argument passed to the operator
     """
     __slots__ = ()