+++ pyccel/ast/core.py
@@ -15,7 +15,7 @@ from pyccel.errors.messages   import RECURSIVE_RESULTS_REQUIRED
 from pyccel.utilities.stage   import PyccelStage
 from pyccel.utilities.strings import create_incremented_string
 
-from .basic     import Basic, PyccelAstNode, iterable, ScopedNode
+from .basic     import PyccelAstNode, TypedAstNode, iterable, ScopedAstNode
 from .builtins  import (PythonEnumerate, PythonLen, PythonMap, PythonTuple,
                         PythonRange, PythonZip, PythonBool, Lambda)
 from .datatypes import (datatype, DataType, NativeSymbol,
@@ -30,7 +30,7 @@ from .operators import PyccelAdd, PyccelMinus, PyccelMul, PyccelDiv, PyccelMod, 
 from .operators import PyccelOperator, PyccelAssociativeParenthesis, PyccelIs
 
 from .variable import DottedName, IndexedElement
-from .variable import Variable
+from .variable import Variable, AnnotatedPyccelSymbol
 
 errors = Errors()
 pyccel_stage = PyccelStage()
@@ -161,7 +161,7 @@ def subs(expr, new_elements):
         new_expr = expr.subs(new_elements)
         new_expr.set_fst(expr.fst)
         return new_expr
-    elif isinstance(expr, PyccelAstNode):
+    elif isinstance(expr, TypedAstNode):
         return expr.subs(new_elements)
 
     else:
@@ -205,7 +205,7 @@ def create_variable(forbidden_names, prefix = None, counter = 1):
     return PyccelSymbol(name, is_temp=True), counter
 
 
-class AsName(Basic):
+class AsName(PyccelAstNode):
 
     """
     Represents a renaming of a variable, used with Import.
@@ -222,7 +222,7 @@ class AsName(Basic):
 
     Parameters
     ==========
-    obj    : Basic or BasicType
+    obj    : PyccelAstNode or PyccelAstNodeType
              The variable, function, or module being renamed
     target : str
              name of variable or function in this context
@@ -232,9 +232,9 @@ class AsName(Basic):
 
     def __init__(self, obj, target):
         if pyccel_stage != "syntactic":
-            assert (isinstance(obj, Basic) and \
+            assert (isinstance(obj, PyccelAstNode) and \
                     not isinstance(obj, PyccelSymbol)) or \
-                   (isinstance(obj, type) and issubclass(obj, Basic))
+                   (isinstance(obj, type) and issubclass(obj, PyccelAstNode))
         self._obj = obj
         self._target = target
         super().__init__()
@@ -282,13 +282,13 @@ class AsName(Basic):
         return hash(self.target)
 
 
-class Duplicate(PyccelAstNode):
+class Duplicate(TypedAstNode):
 
     """ this is equivalent to the * operator for python tuples.
 
     Parameters
     ----------
-    value : PyccelAstNode
+    value : TypedAstNode
            an expression which represents the initilized value of the list
 
     shape : the shape of the array
@@ -321,13 +321,13 @@ class Duplicate(PyccelAstNode):
     def __repr__(self):
         return '{} * {}'.format(repr(self.val), repr(self.length))
 
-class Concatenate(PyccelAstNode):
+class Concatenate(TypedAstNode):
 
     """ this is equivalent to the + operator for python tuples
 
     Parameters
     ----------
-    args : PyccelAstNodes
+    args : TypedAstNodes
            The tuples
     """
     __slots__ = ('_args','_dtype','_precision','_rank','_shape','_order')
@@ -349,33 +349,39 @@ class Concatenate(PyccelAstNode):
         return self._args
 
 
-class Assign(Basic):
+class Assign(PyccelAstNode):
+    """
+    Represents variable assignment for code generation.
 
-    """Represents variable assignment for code generation.
+    Class representing an assignment node, where the result of an expression
+    (rhs: right hand side) is saved into a variable (lhs: left hand side).
 
     Parameters
     ----------
-    lhs : PyccelAstNode
+    lhs : TypedAstNode
         In the syntactic stage:
            Object representing the lhs of the expression. These should be
            singular objects, such as one would use in writing code. Notable types
            include PyccelSymbol, and IndexedElement. Types that
            subclass these types are also supported.
         In the semantic stage:
-           Variable or IndexedElement
+           Variable or IndexedElement.
 
-    rhs : PyccelAstNode
+    rhs : TypedAstNode
         In the syntactic stage:
-          Object representing the rhs of the expression
+          Object representing the rhs of the expression.
         In the semantic stage :
-          PyccelAstNode with the same shape as the lhs
+          TypedAstNode with the same shape as the lhs.
 
-    status: None, str
-        if lhs is not allocatable, then status is None.
-        otherwise, status is {'allocated', 'unallocated'}
+    status : str, optional
+        If lhs is not allocatable, then status is None.
+        otherwise, status is {'allocated', 'unallocated'}.
 
-    like: None, Variable
-        contains the name of the variable from which the lhs will be cloned.
+    like : Variable, optional
+        Contains the name of the variable from which the lhs will be cloned.
+
+    fst : ast.Ast
+        The ast object parsed by Python's ast module.
 
     Examples
     --------
@@ -480,7 +486,7 @@ class Assign(Basic):
         return False
 
 #------------------------------------------------------------------------------
-class Allocate(Basic):
+class Allocate(PyccelAstNode):
     """
     Represents memory allocation for code generation.
 
@@ -577,7 +583,7 @@ class Allocate(Basic):
         return hash((id(self.variable), self.shape, self.order, self.status))
 
 #------------------------------------------------------------------------------
-class Deallocate(Basic):
+class Deallocate(PyccelAstNode):
     """
     Represents memory deallocation (usually of an array) for code generation.
     This is relevant to low-level target languages, such as C or Fortran,
@@ -622,7 +628,7 @@ class Deallocate(Basic):
         return hash(id(self.variable))
 
 #------------------------------------------------------------------------------
-class CodeBlock(Basic):
+class CodeBlock(PyccelAstNode):
 
     """Represents a list of stmt for code generation.
        we use it when a single statement in python
@@ -705,7 +711,7 @@ class CodeBlock(Basic):
             if not l.fst:
                 l.set_fst(fst)
 
-class AliasAssign(Basic):
+class AliasAssign(PyccelAstNode):
 
     """Represents aliasing for code generation. An alias is any statement of the
     form `lhs := rhs` where
@@ -758,7 +764,7 @@ class AliasAssign(Basic):
         return self._rhs
 
 
-class SymbolicAssign(Basic):
+class SymbolicAssign(PyccelAstNode):
 
     """Represents symbolic aliasing for code generation. An alias is any statement of the
     form `lhs := rhs` where
@@ -805,7 +811,7 @@ class AugAssign(Assign):
 
     Parameters
     ----------
-    lhs : PyccelAstNode
+    lhs : TypedAstNode
         In the syntactic stage:
            Object representing the lhs of the expression. These should be
            singular objects, such as one would use in writing code. Notable types
@@ -817,11 +823,11 @@ class AugAssign(Assign):
     op : str
         Operator (+, -, /, \*, %).
 
-    rhs : PyccelAstNode
+    rhs : TypedAstNode
         In the syntactic stage:
           Object representing the rhs of the expression
         In the semantic stage :
-          PyccelAstNode with the same shape as the lhs
+          TypedAstNode with the same shape as the lhs
 
     status: None, str
         if lhs is not allocatable, then status is None.
@@ -889,7 +895,7 @@ class AugAssign(Assign):
                 like   = self.like)
 
 
-class While(ScopedNode):
+class While(ScopedAstNode):
 
     """Represents a 'while' statement in the code.
 
@@ -899,7 +905,7 @@ class While(ScopedNode):
 
     Parameters
     ----------
-    test : PyccelAstNode
+    test : TypedAstNode
         test condition given as an expression
     body : list of Pyccel objects
         list of statements representing the body of the While statement.
@@ -946,7 +952,7 @@ class While(ScopedNode):
         return tuple(self.scope.variables.values())
 
 
-class With(ScopedNode):
+class With(ScopedAstNode):
 
     """Represents a 'with' statement in the code.
 
@@ -956,7 +962,7 @@ class With(ScopedNode):
 
     Parameters
     ----------
-    test : PyccelAstNode
+    test : TypedAstNode
         with definition statement given as an expression
     body : list of Pyccel objects
         list of statements representing the body of the With statement.
@@ -1015,7 +1021,7 @@ class With(ScopedNode):
 
 # TODO add a name to a block?
 
-class Block(ScopedNode):
+class Block(ScopedAstNode):
 
     """Represents a block in the code. A block consists of the following inputs
 
@@ -1081,7 +1087,7 @@ class Block(ScopedNode):
 
 
 
-class Module(ScopedNode):
+class Module(ScopedAstNode):
     """
     Represents a module in the code.
 
@@ -1351,7 +1357,7 @@ class Module(ScopedNode):
         """
         return self._internal_dictionary.keys()
 
-class ModuleHeader(Basic):
+class ModuleHeader(PyccelAstNode):
     """
     Represents the header file for a module.
 
@@ -1407,7 +1413,7 @@ class ModuleHeader(Basic):
     def module(self):
         return self._module
 
-class Program(ScopedNode):
+class Program(ScopedAstNode):
 
     """Represents a Program in the code. A block consists of the following inputs
 
@@ -1494,7 +1500,7 @@ class Program(ScopedNode):
 
 
 #==============================================================================
-class Iterable(Basic):
+class Iterable(PyccelAstNode):
     """
     Wrapper around iterable types helping to convert between those
     types and a range (necessary in low level languages, e.g. C and Fortran)
@@ -1620,7 +1626,7 @@ class Iterable(Basic):
 
 #==============================================================================
 
-class For(ScopedNode):
+class For(ScopedAstNode):
 
     """Represents a 'for-loop' in the code.
 
@@ -1767,24 +1773,31 @@ class ForIterator(For):
     def ranges(self):
         return get_iterable_ranges(self.iterable)
 
-class FunctionCallArgument(Basic):
+class FunctionCallArgument(PyccelAstNode):
     """
-    An argument passed in a function call
+    An argument passed in a function call.
+
+    Class describing an argument passed to a function in a
+    function call.
 
     Parameters
-    ---------
-    value   : PyccelAstNode
-              The expression passed as an argument
-    keyword : str
-              If the argument is passed by keyword then this
-              is that keyword
+    ----------
+    value : TypedAstNode
+        The expression passed as an argument.
+    keyword : str, optional
+        If the argument is passed by keyword then this
+        is that keyword.
+    fst : ast.Ast
+        The ast object parsed by Python's ast module.
     """
     __slots__ = ('_value', '_keyword')
     _attribute_nodes = ('_value',)
-    def __init__(self, value, keyword = None):
+    def __init__(self, value, keyword = None, *, fst = None):
         self._value = value
         self._keyword = keyword
         super().__init__()
+        if fst:
+            self.set_fst(fst)
 
     @property
     def value(self):
@@ -1816,7 +1829,7 @@ class FunctionCallArgument(Basic):
         else:
             return '{}'.format(str(self.value))
 
-class FunctionDefArgument(PyccelAstNode):
+class FunctionDefArgument(TypedAstNode):
     """
     Node describing the argument of a function.
 
@@ -1829,7 +1842,7 @@ class FunctionDefArgument(PyccelAstNode):
     name : PyccelSymbol, Variable, FunctionAddress
         The name of the argument.
 
-    value : PyccelAstNode, default: None
+    value : TypedAstNode, default: None
         The default value of the argument.
 
     kwonly : bool
@@ -1859,6 +1872,9 @@ class FunctionDefArgument(PyccelAstNode):
         elif isinstance(name, PyccelSymbol):
             self._var  = name
             self._name = name
+        elif isinstance(name, AnnotatedPyccelSymbol):
+            self._var  = name
+            self._name = name.name
         else:
             raise TypeError("Name must be a PyccelSymbol, Variable or FunctionAddress")
         self._value      = value
@@ -1895,7 +1911,10 @@ class FunctionDefArgument(PyccelAstNode):
 
     @property
     def annotation(self):
-        """ The argument annotation providing dtype information
+        """
+        The argument annotation providing dtype information.
+
+        The argument annotation providing dtype information.
         """
         return self._annotation
 
@@ -1955,7 +1974,7 @@ class FunctionDefArgument(PyccelAstNode):
         else:
             return 'FunctionDefArgument({})'.format(repr(self.name))
 
-class FunctionDefResult(PyccelAstNode):
+class FunctionDefResult(TypedAstNode):
     """
     Node describing the result of a function.
 
@@ -1990,10 +2009,10 @@ class FunctionDefResult(PyccelAstNode):
         self._annotation = annotation
 
         if pyccel_stage == 'syntactic':
-            if not isinstance(var, PyccelSymbol):
-                raise TypeError("Var must be a PyccelSymbol")
+            if not isinstance(var, (PyccelSymbol, AnnotatedPyccelSymbol)):
+                raise TypeError(f"Var must be a PyccelSymbol or an AnnotatedPyccelSymbol, not a {type(var)}")
         elif not isinstance(var, Variable):
-            raise TypeError("Var must be a Variable")
+            raise TypeError(f"Var must be a Variable not a {type(var)}")
         else:
             self._is_argument = var.is_argument
 
@@ -2036,7 +2055,7 @@ class FunctionDefResult(PyccelAstNode):
     def __str__(self):
         return str(self.var)
 
-class FunctionCall(PyccelAstNode):
+class FunctionCall(TypedAstNode):
     """
     Represents a function call in the code.
 
@@ -2117,11 +2136,11 @@ class FunctionCall(PyccelAstNode):
 
         # Handle function as argument
         arg_vals = [None if a is None else a.value for a in args]
-        args = [FunctionCallArgument(FunctionAddress(av.name, av.arguments, av.results, []), keyword=a.keyword)
+        args = [FunctionCallArgument(FunctionAddress(av.name, av.arguments, av.results), keyword=a.keyword)
                 if isinstance(av, FunctionDef) else a for a, av in zip(args, arg_vals)]
 
         if current_function == func.name:
-            if len(func.results)>0 and not isinstance(func.results[0], PyccelAstNode):
+            if len(func.results)>0 and not isinstance(func.results[0], TypedAstNode):
                 errors.report(RECURSIVE_RESULTS_REQUIRED, symbol=func, severity="fatal")
 
         self._funcdef       = func
@@ -2188,7 +2207,7 @@ class DottedFunctionCall(FunctionCall):
                        The definition of the function being called
     args             : tuple
                        The arguments being passed to the function
-    prefix           : PyccelAstNode
+    prefix           : TypedAstNode
                        The object in which the function is defined
                        E.g. for a.f()
                        prefix will contain a
@@ -2267,13 +2286,13 @@ class ConstructorCall(DottedFunctionCall):
         return self._cls_variable
 
 
-class Return(Basic):
+class Return(PyccelAstNode):
 
     """Represents a function return in the code.
 
     Parameters
     ----------
-    expr : PyccelAstNode
+    expr : TypedAstNode
         The expression to return.
 
     stmts :represent assign stmts in the case of expression return
@@ -2312,7 +2331,7 @@ class Return(Basic):
             code = ''
         return code+"Return({})".format(','.join([repr(e) for e in self.expr]))
 
-class FunctionDef(ScopedNode):
+class FunctionDef(ScopedAstNode):
 
     """
     Represents a function definition.
@@ -2940,7 +2959,7 @@ class PyccelFunctionDef(FunctionDef):
     name : str
         The name of the function.
 
-    func_class : type inheriting from PyccelInternalFunction / PyccelAstNode
+    func_class : type inheriting from PyccelInternalFunction / TypedAstNode
         The class which should be instantiated upon a FunctionCall
         to this FunctionDef object.
 
@@ -2956,7 +2975,7 @@ class PyccelFunctionDef(FunctionDef):
     __slots__ = ('_argument_description',)
     def __init__(self, name, func_class, *, decorators = {}, argument_description = {}):
         assert isinstance(func_class, type) and \
-                issubclass(func_class, (PyccelInternalFunction, PyccelAstNode))
+                issubclass(func_class, (PyccelInternalFunction, TypedAstNode))
         assert isinstance(argument_description, dict)
         arguments = ()
         results = ()
@@ -2976,7 +2995,7 @@ class PyccelFunctionDef(FunctionDef):
         """
         return self._argument_description
 
-class Interface(Basic):
+class Interface(PyccelAstNode):
 
     """Represents an Interface.
 
@@ -3073,8 +3092,10 @@ class Interface(Basic):
                         severity='fatal')
 
 class FunctionAddress(FunctionDef):
+    """
+    Represents a function address.
 
-    """Represents a function address.
+    A function definition can have a FunctionAddress as an argument.
 
     Parameters
     ----------
@@ -3087,31 +3108,35 @@ class FunctionAddress(FunctionDef):
     results : iterable
         The direct outputs of the function address.
 
-    is_argument: bool
-        if object is the argument of a function [Default value: False]
+    is_optional : bool
+        If object is an optional argument of a function [Default value: False].
 
-    is_kwonly: bool
-        if object is an argument which can only be specified using its keyword
+    is_kwonly : bool
+        If object is an argument which can only be specified using its keyword.
 
-    is_optional: bool
-        if object is an optional argument of a function [Default value: False]
+    is_argument : bool
+        If object is the argument of a function [Default value: False].
 
-    memory_handling: str
-        must be 'heap', 'stack' or 'alias' [Default value: 'stack']
+    memory_handling : str
+        Must be 'heap', 'stack' or 'alias' [Default value: 'stack'].
+
+    **kwargs : dict
+        Any keyword arguments which should be passed to the super class FunctionDef.
+
+    See Also
+    --------
+    FunctionDef
+        The super class from which this object derives.
 
     Examples
     --------
     >>> from pyccel.ast.core import Variable, FunctionAddress, FuncAddressDeclare, FunctionDef
     >>> x = Variable('float', 'x')
     >>> y = Variable('float', 'y')
-
-    a function definition can have a FunctionAddress as an argument
-
-    >>> FunctionDef('g', [FunctionAddress('f', [x], [y], [])], [], [])
-
-    we can also Declare a FunctionAddress
-
-    >>> FuncAddressDeclare(FunctionAddress('f', [x], [y], []))
+    >>> # a function definition can have a FunctionAddress as an argument
+    >>> FunctionDef('g', [FunctionAddress('f', [x], [y])], [], [])
+    >>> # we can also Declare a FunctionAddress
+    >>> FuncAddressDeclare(FunctionAddress('f', [x], [y]))
     """
     __slots__ = ('_is_optional','_is_kwonly','_is_argument', '_memory_handling')
 
@@ -3120,14 +3145,13 @@ class FunctionAddress(FunctionDef):
         name,
         arguments,
         results,
-        body,
         is_optional=False,
         is_kwonly=False,
         is_argument=False,
         memory_handling='stack',
         **kwargs
         ):
-        super().__init__(name, arguments, results, body, scope=1,**kwargs)
+        super().__init__(name, arguments, results, body=[], scope=None, **kwargs)
         if not isinstance(is_argument, bool):
             raise TypeError('Expecting a boolean for is_argument')
 
@@ -3173,6 +3197,28 @@ class FunctionAddress(FunctionDef):
     def is_optional(self):
         return self._is_optional
 
+    def __getnewargs__(self):
+        """
+        Function called during unpickling.
+
+        For more details see : https://docs.python.org/3/library/pickle.html#object.__getnewargs__.
+
+        Returns
+        -------
+        args
+            A tuple containing any arguments to be passed to the constructor.
+        kwargs
+            A dict containing any keyword arguments to be passed to the constructor.
+        """
+        args, kwargs = super().__getnewargs__()
+        args = args[:-1] # Remove body argument
+        kwargs.pop('scope')
+        kwargs['is_argument'] = self.is_argument
+        kwargs['is_kwonly'] = self.is_kwonly
+        kwargs['is_optional'] = self.is_optional
+        kwargs['memory_handling'] = self.memory_handling
+        return args, kwargs
+
 class SympyFunction(FunctionDef):
 
     """Represents a function definition."""
@@ -3180,7 +3226,7 @@ class SympyFunction(FunctionDef):
 
 
 
-class ClassDef(ScopedNode):
+class ClassDef(ScopedAstNode):
     """
     Represents a class definition.
 
@@ -3554,7 +3600,7 @@ class ClassDef(ScopedNode):
         return False
 
 
-class Import(Basic):
+class Import(PyccelAstNode):
 
     """Represents inclusion of dependencies in the code.
 
@@ -3690,19 +3736,21 @@ class Import(Basic):
 
 # TODO: Should Declare have an optional init value for each var?
 
-class FuncAddressDeclare(Basic):
+class FuncAddressDeclare(PyccelAstNode):
+    """
+    Represents a FunctionAddress declaration in the code.
 
-    """Represents a FunctionAddress declaration in the code.
+    Represents a FunctionAddress declaration in the code.
 
     Parameters
     ----------
-    variable:
+    variable : FunctionAddress
         An instance of FunctionAddress.
-    intent: None, str
-        one among {'in', 'out', 'inout'}
-    value: PyccelAstNode
-        variable value
-    static: bool
+    intent : str, optional
+        One among {'in', 'out', 'inout'}.
+    value : TypedAstNode
+        Variable value.
+    static : bool
         True for a static declaration of an array.
 
     Examples
@@ -3710,7 +3758,7 @@ class FuncAddressDeclare(Basic):
     >>> from pyccel.ast.core import Variable, FunctionAddress, FuncAddressDeclare
     >>> x = Variable('float', 'x')
     >>> y = Variable('float', 'y')
-    >>> FuncAddressDeclare(FunctionAddress('f', [x], [y], []))
+    >>> FuncAddressDeclare(FunctionAddress('f', [x], [y]))
     """
     __slots__ = ('_variable','_intent','_value','_static')
     _attribute_nodes = ('_variable', '_value')
@@ -3768,7 +3816,7 @@ class FuncAddressDeclare(Basic):
         return self._static
 
 # ARA : issue-999 add is_external for external function exported through header files
-class Declare(Basic):
+class Declare(PyccelAstNode):
 
     """Represents a variable declaration in the code.
 
@@ -3781,7 +3829,7 @@ class Declare(Basic):
         Variables must be of the same type.
     intent: None, str
         one among {'in', 'out', 'inout'}
-    value: PyccelAstNode
+    value: TypedAstNode
         variable value
     static: bool
         True for a static declaration of an array.
@@ -3890,21 +3938,21 @@ class Declare(Basic):
     def __repr__(self):
         return 'Declare({})'.format(repr(self.variable))
 
-class Break(Basic):
+class Break(PyccelAstNode):
 
     """Represents a break in the code."""
     __slots__ = ()
     _attribute_nodes = ()
 
 
-class Continue(Basic):
+class Continue(PyccelAstNode):
 
     """Represents a continue in the code."""
     __slots__ = ()
     _attribute_nodes = ()
 
 
-class Raise(Basic):
+class Raise(PyccelAstNode):
 
     """Represents a raise in the code."""
     __slots__ = ()
@@ -3912,13 +3960,13 @@ class Raise(Basic):
 
 
 
-class SymbolicPrint(Basic):
+class SymbolicPrint(PyccelAstNode):
 
     """Represents a print function of symbolic expressions in the code.
 
     Parameters
     ----------
-    expr : PyccelAstNode
+    expr : TypedAstNode
         The expression to print
 
     Examples
@@ -3950,7 +3998,7 @@ class SymbolicPrint(Basic):
         return self._expr
 
 
-class Del(Basic):
+class Del(PyccelAstNode):
 
     """Represents a memory deallocation in the code.
 
@@ -3984,7 +4032,7 @@ class Del(Basic):
         return self._variables
 
 
-class EmptyNode(Basic):
+class EmptyNode(PyccelAstNode):
     """
     Represents an empty node in the abstract syntax tree (AST).
     When a subtree is removed from the AST, we replace it with an EmptyNode
@@ -4010,7 +4058,7 @@ class EmptyNode(Basic):
         return ''
 
 
-class Comment(Basic):
+class Comment(PyccelAstNode):
 
     """Represents a Comment in the code.
 
@@ -4082,7 +4130,7 @@ class SeparatorComment(Comment):
         text = """.""" * n
         super().__init__(text)
 
-class AnnotatedComment(Basic):
+class AnnotatedComment(PyccelAstNode):
 
     """Represents a Annotated Comment in the code.
 
@@ -4122,7 +4170,7 @@ class AnnotatedComment(Basic):
         args = (self.accel, self.txt)
         return args
 
-class CommentBlock(Basic):
+class CommentBlock(PyccelAstNode):
 
     """ Represents a Block of Comments
 
@@ -4158,13 +4206,13 @@ class CommentBlock(Basic):
         self._header = header
 
 
-class Assert(Basic):
+class Assert(PyccelAstNode):
     """
     Represents an assert statement in the code.
 
     Parameters
     ----------
-    test: PyccelAstNode
+    test: TypedAstNode
         boolean expression to check
     """
     __slots__ = ('_test',)
@@ -4182,13 +4230,13 @@ class Assert(Basic):
         return self._test
 
 
-class Pass(Basic):
+class Pass(PyccelAstNode):
 
     """Basic class for pass instruction."""
     __slots__ = ()
     _attribute_nodes = ()
 
-class Exit(Basic):
+class Exit(PyccelAstNode):
 
     """Basic class for exits."""
     __slots__ = ()
@@ -4200,13 +4248,13 @@ class ErrorExit(Exit):
     """Exit with error."""
     __slots__ = ()
 
-class IfSection(Basic):
+class IfSection(PyccelAstNode):
     """Represents a condition and associated code block
     in an if statement in the code.
 
     Parameters
     ----------
-    cond : PyccelAstNode
+    cond : TypedAstNode
            A boolean expression indicating whether or not the block
            should be executed
     body : CodeBlock
@@ -4253,7 +4301,7 @@ class IfSection(Basic):
     def __str__(self):
         return "IfSec({},{})".format(str(self.condition), str(self.body))
 
-class If(Basic):
+class If(PyccelAstNode):
 
     """Represents a if statement in the code.
 
@@ -4297,7 +4345,7 @@ class If(Basic):
     def __str__(self):
         return "If({})".format(','.join(str(b) for b in self.blocks))
 
-class StarredArguments(Basic):
+class StarredArguments(PyccelAstNode):
     __slots__ = ('_starred_obj',)
     _attribute_nodes = ('_starred_obj',)
     def __init__(self, args):
@@ -4310,7 +4358,7 @@ class StarredArguments(Basic):
 
 # ...
 
-class InProgram(PyccelAstNode):
+class InProgram(TypedAstNode):
     """
     Class representing the boolean:
     __name__ == '__main__'
@@ -4325,7 +4373,7 @@ class InProgram(PyccelAstNode):
 
 # ...
 
-class Decorator(Basic):
+class Decorator(PyccelAstNode):
     """ Class representing a function decorator.
     For now this is just designed to handle the pyccel decorators