+++ pyccel/ast/variable.py
@@ -12,11 +12,11 @@ import inspect
 from pyccel.errors.errors   import Errors
 from pyccel.utilities.stage import PyccelStage
 
-from .basic     import Basic, PyccelAstNode
+from .basic     import PyccelAstNode, TypedAstNode
 from .datatypes import (datatype, DataType,
                         NativeInteger, NativeBool, NativeFloat,
                         NativeComplex)
-from .internals import PyccelArrayShapeElement, Slice, get_final_precision
+from .internals import PyccelArrayShapeElement, Slice, get_final_precision, PyccelSymbol
 from .literals  import LiteralInteger, Nil
 from .operators import (PyccelMinus, PyccelDiv, PyccelMul,
                         PyccelUnarySub, PyccelAdd)
@@ -25,6 +25,7 @@ errors = Errors()
 pyccel_stage = PyccelStage()
 
 __all__ = (
+    'AnnotatedPyccelSymbol',
     'Constant',
     'DottedName',
     'DottedVariable',
@@ -35,7 +36,7 @@ __all__ = (
     'Variable'
 )
 
-class Variable(PyccelAstNode):
+class Variable(TypedAstNode):
     """
     Represents a typed variable.
 
@@ -74,7 +75,7 @@ class Variable(PyccelAstNode):
 
     shape : tuple, default: None
         The shape of the array. A tuple whose elements indicate the number of elements along
-        each of the dimensions of an array. The elements of the tuple should be None or PyccelAstNodes.
+        each of the dimensions of an array. The elements of the tuple should be None or TypedAstNodes.
 
     cls_base : class, default: None
         Class base if variable is an object or an object member.
@@ -180,7 +181,7 @@ class Variable(PyccelAstNode):
         self._is_argument    = is_argument
         self._is_temp        = is_temp
 
-        # ------------ PyccelAstNode Properties ---------------
+        # ------------ TypedAstNode Properties ---------------
         if isinstance(dtype, str) or str(dtype) == '*':
 
             dtype = datatype(dtype)
@@ -252,11 +253,11 @@ class Variable(PyccelAstNode):
                 new_shape.append(s)
             elif isinstance(s, int):
                 new_shape.append(LiteralInteger(s))
-            elif s is None or isinstance(s, PyccelAstNode):
+            elif s is None or isinstance(s, TypedAstNode):
                 new_shape.append(PyccelArrayShapeElement(self, LiteralInteger(i)))
             else:
                 raise TypeError('shape elements cannot be '+str(type(s))+'. They must be one of the following types: LiteralInteger,'
-                                'Variable, Slice, PyccelAstNode, int, Function')
+                                'Variable, Slice, TypedAstNode, int, Function')
         return tuple(new_shape)
 
     def shape_can_change(self, i):
@@ -299,7 +300,7 @@ class Variable(PyccelAstNode):
         ----------
         shape : tuple
             The shape of the array. A tuple whose elements indicate the number of elements along
-            each of the dimensions of an array. The elements of the tuple should be None or PyccelAstNodes.
+            each of the dimensions of an array. The elements of the tuple should be None or TypedAstNodes.
         """
         self._alloc_shape = shape
         self._shape = self.process_shape(shape)
@@ -370,8 +371,11 @@ class Variable(PyccelAstNode):
 
     @property
     def is_const(self):
-        """ Indicates if the Variable is constant
-        within its context
+        """
+        Indicates whether the Variable is constant within its context.
+
+        Indicates whether the Variable is constant within its context.
+        True if the Variable is constant, false if it can be modified.
         """
         return self._is_const
 
@@ -588,7 +592,7 @@ class Variable(PyccelAstNode):
             raise ValueError("Variables cannot become temporary")
         self._is_temp = is_temp
 
-class DottedName(Basic):
+class DottedName(PyccelAstNode):
 
     """
     Represents a dotted object.
@@ -611,6 +615,12 @@ class DottedName(Basic):
     """
     __slots__ = ('_name',)
     _attribute_nodes = ()
+
+    def __new__(cls, *args):
+        if len(args) == 1:
+            return args[0]
+        else:
+            return super().__new__(cls)
 
     def __init__(self, *args):
 
@@ -834,10 +844,29 @@ class Constant(Variable):
 
 
 
-class IndexedElement(PyccelAstNode):
-
+class IndexedElement(TypedAstNode):
     """
-    Represents a mathematical object with indices.
+    Represents an indexed object in the code.
+
+    Represents an object which is a subset of a base object. The
+    indexed object is retrieved by passing indices to the base
+    object using the `[]` syntax.
+
+    In the semantic stage, the base object is an array, tuple or
+    list. This function then determines the new rank and shape of
+    the data block.
+
+    In the syntactic stage, this object is more versatile, it
+    stores anything which is indexed using `[]` syntax. This can
+    additionally include classes, maps, etc.
+
+    Parameters
+    ----------
+    base : Variable | PyccelSymbol | DottedName
+        The object being indexed.
+
+    *indices : tuple of TypedAstNode
+        The values used to index the base.
 
     Examples
     --------
@@ -845,7 +874,7 @@ class IndexedElement(PyccelAstNode):
     >>> A = Variable('A', dtype='int', shape=(2,3), rank=2)
     >>> i = Variable('i', dtype='int')
     >>> j = Variable('j', dtype='int')
-    >>> IndexedElement(A, i, j)
+    >>> IndexedElement(A, (i, j))
     IndexedElement(A, i, j)
     >>> IndexedElement(A, i, j) == A[i, j]
     True
@@ -853,20 +882,15 @@ class IndexedElement(PyccelAstNode):
     __slots__ = ('_label', '_indices','_dtype','_precision','_shape','_rank','_order')
     _attribute_nodes = ('_label', '_indices')
 
-    def __init__(
-        self,
-        base,
-        *args,
-        **kw_args
-        ):
+    def __init__(self, base, *indices):
 
-        if not args:
+        if not indices:
             raise IndexError('Indexed needs at least one index.')
 
         self._label = base
 
         if pyccel_stage == 'syntactic':
-            self._indices = args
+            self._indices = indices
             super().__init__()
             return
 
@@ -877,48 +901,40 @@ class IndexedElement(PyccelAstNode):
         rank  = base.rank
 
         # Add empty slices to fully index the object
-        if len(args) < rank:
-            args = args + tuple([Slice(None, None)]*(rank-len(args)))
+        if len(indices) < rank:
+            indices = indices + tuple([Slice(None, None)]*(rank-len(indices)))
 
-        if any(not isinstance(a, (int, PyccelAstNode, Slice)) for a in args):
+        if any(not isinstance(a, (int, TypedAstNode, Slice)) for a in indices):
             errors.report("Index is not of valid type",
-                    symbol = args, severity = 'fatal')
+                    symbol = indices, severity = 'fatal')
 
-        self._indices = tuple(LiteralInteger(a) if isinstance(a, int) else a for a in args)
+        self._indices = tuple(LiteralInteger(a) if isinstance(a, int) else a for a in indices)
         super().__init__()
 
         # Calculate new shape
+        new_shape = []
+        from .mathext import MathCeil
+        for a,s in zip(indices, shape):
+            if isinstance(a, Slice):
+                start = a.start
+                stop  = a.stop if a.stop is not None else s
+                step  = a.step
+                if isinstance(start, PyccelUnarySub):
+                    start = PyccelAdd(s, start, simplify=True)
+                if isinstance(stop, PyccelUnarySub):
+                    stop = PyccelAdd(s, stop, simplify=True)
 
-        if shape is not None:
-            new_shape = []
-            from .mathext import MathCeil
-            for a,s in zip(args, shape):
-                if isinstance(a, Slice):
-                    start = a.start
-                    stop  = a.stop if a.stop is not None else s
-                    step  = a.step
-                    if isinstance(start, PyccelUnarySub):
-                        start = PyccelAdd(s, start, simplify=True)
-                    if isinstance(stop, PyccelUnarySub):
-                        stop = PyccelAdd(s, stop, simplify=True)
+                _shape = stop if start is None else PyccelMinus(stop, start, simplify=True)
+                if step is not None:
+                    if isinstance(step, PyccelUnarySub):
+                        start = s if a.start is None else start
+                        _shape = start if a.stop is None else PyccelMinus(start, stop, simplify=True)
+                        step = PyccelUnarySub(step)
 
-                    _shape = stop if start is None else PyccelMinus(stop, start, simplify=True)
-                    if step is not None:
-                        if isinstance(step, PyccelUnarySub):
-                            start = s if a.start is None else start
-                            _shape = start if a.stop is None else PyccelMinus(start, stop, simplify=True)
-                            step = PyccelUnarySub(step)
-
-                        _shape = MathCeil(PyccelDiv(_shape, step, simplify=True))
-                    new_shape.append(_shape)
-            self._rank  = len(new_shape)
-            self._shape = None if self._rank == 0 else tuple(new_shape)
-        else:
-            new_rank = rank
-            for i in range(rank):
-                if not isinstance(args[i], Slice):
-                    new_rank -= 1
-            self._rank = new_rank
+                    _shape = MathCeil(PyccelDiv(_shape, step, simplify=True))
+                new_shape.append(_shape)
+        self._rank  = len(new_shape)
+        self._shape = None if self._rank == 0 else tuple(new_shape)
 
         self._order = None if self.rank < 2 else base.order
 
@@ -963,6 +979,16 @@ class IndexedElement(PyccelAstNode):
                 j += 1
             new_indexes.append(i)
         return IndexedElement(base, *new_indexes)
+
+    @property
+    def is_const(self):
+        """
+        Indicates whether the Variable is constant within its context.
+
+        Indicates whether the Variable is constant within its context.
+        True if the Variable is constant, false if it can be modified.
+        """
+        return self.base.is_const
 
 class DottedVariable(Variable):
 
@@ -1013,3 +1039,58 @@ class DottedVariable(Variable):
         dtype = repr(self.dtype)
         classname = type(self).__name__
         return f'{classname}({lhs}.{name}, dtype={dtype}'
+
+class AnnotatedPyccelSymbol(PyccelAstNode):
+    """
+    Class representing a symbol in the code which has an annotation.
+
+    Symbolic placeholder for a Python variable, which has a name but no type yet.
+    This is very generic, and it can also represent a function or a module.
+
+    Parameters
+    ----------
+    name : str
+        Name of the symbol.
+
+    annotation : SyntacticTypeAnnotation
+        The annotation describing the type that the object will have.
+
+    is_temp : bool
+        Indicates if the symbol is a temporary object. This either means that the
+        symbol represents an object originally named `_` in the code, or that the
+        symbol represents an object created by Pyccel in order to assign a
+        temporary object. This is sometimes necessary to facilitate the translation.
+    """
+    __slots__ = ('_name', '_annotation')
+    _attribute_nodes = ()
+
+    def __init__(self, name, annotation, is_temp = False):
+        if isinstance(name, (PyccelSymbol, DottedName)):
+            self._name = name
+        elif isinstance(name, str):
+            self._name = PyccelSymbol(name, is_temp)
+        else:
+            raise TypeError(f"Name should be a string or a PyccelSymbol not a {type(name)}")
+        self._annotation = annotation
+        super().__init__()
+
+    @property
+    def name(self):
+        """
+        Get the PyccelSymbol describing the name.
+
+        Get the PyccelSymbol describing the name of the symbol in the code.
+        """
+        return self._name
+
+    @property
+    def annotation(self):
+        """
+        Get the annotation.
+
+        Get the annotation left on the symbol. This should be a type annotation.
+        """
+        return self._annotation
+
+    def __str__(self):
+        return f'{self.name} : {self.annotation}'